refactor(contributors): split for better maintainability
