+++ frontend/src/constants/accessColors.ts
@@ -1,0 +1,4 @@
+export const ACCESS_COLORS: Record<string, string> = {
+  VIEW: "bg-blue-500",
+  EDIT: "bg-green-500",
+};
