+++ frontend/src/constants/constants.ts
@@ -1,0 +1,1 @@
+export { ACCESS_COLORS } from "./accessColors";
