+++ frontend/src/hooks/hooks.ts
@@ -9,3 +9,4 @@ export { useUploadSubtitle } from "./subtitles/useUploadSubtitle";
 export { useDeleteSubtitleDoc } from "./subtitles/useDeleteSubtitleDoc";
 export { useShareSubtitleDoc } from "./subtitles/useShareSubtitleDoc";
 export { useRemoveAccess } from "./subtitles/useRemoveAccess";
+export { useContributorActions } from "./contributors/useContributorActions";
